Import test fixes

+++ pkg/ccl/importccl/import_stmt_test.go
@@ -495,8 +495,8 @@ COPY t (a, b, c) FROM stdin;
 				ORDER BY descriptor_name
 				`: {{
 					`CREATE TABLE a (
-	i INT NOT NULL,
-	k INT NULL,
+	i INT8 NOT NULL,
+	k INT8 NULL,
 	CONSTRAINT a_pkey PRIMARY KEY (i ASC),
 	CONSTRAINT a_k_fkey FOREIGN KEY (k) REFERENCES a (i),
 	INDEX a_auto_index_a_k_fkey (k ASC),
@@ -504,7 +504,7 @@ COPY t (a, b, c) FROM stdin;
 	FAMILY "primary" (i, k)
 )`}, {
 					`CREATE TABLE b (
-	j INT NOT NULL,
+	j INT8 NOT NULL,
 	CONSTRAINT b_pkey PRIMARY KEY (j ASC),
 	CONSTRAINT b_j_fkey FOREIGN KEY (j) REFERENCES a (i),
 	FAMILY "primary" (j)
@@ -684,8 +684,8 @@ const (
 )`
 	testPgdumpCreateWeather = `CREATE TABLE weather (
 	city VARCHAR(80) NULL,
-	temp_lo INT NULL,
-	temp_hi INT NULL,
+	temp_lo INT8 NULL,
+	temp_hi INT8 NULL,
 	prcp FLOAT4 NULL,
 	date DATE NULL,
 	CONSTRAINT weather_city_fkey FOREIGN KEY (city) REFERENCES cities (city),
@@ -1329,7 +1329,7 @@ func TestImportCSVStmt(t *testing.T) {
 
 		data = ",5,e,7,,"
 		t.Run(data, func(t *testing.T) {
-			if _, err := conn.Exec(query, srv.URL); !testutils.IsError(err, `row 1: parse "a" as INT: could not parse ""`) {
+			if _, err := conn.Exec(query, srv.URL); !testutils.IsError(err, `row 1: parse "a" as INT8: could not parse ""`) {
 				t.Fatalf("unexpected: %v", err)
 			}
 			if _, err := conn.Exec(query+nullif, srv.URL); !testutils.IsError(err, `row 1: generate insert row: null value in column "a" violates not-null constraint`) {
@@ -2291,7 +2291,7 @@ func TestImportPgDump(t *testing.T) {
 				// Verify table schema because PKs and indexes are at the bottom of pg_dump.
 				sqlDB.CheckQueryResults(t, `SHOW CREATE TABLE simple`, [][]string{{
 					"simple", `CREATE TABLE simple (
-	i INT NOT NULL,
+	i INT8 NOT NULL,
 	s STRING NULL,
 	b BYTES NULL,
 	CONSTRAINT simple_pkey PRIMARY KEY (i ASC),
@@ -2333,7 +2333,7 @@ func TestImportPgDump(t *testing.T) {
 				// Verify table schema because PKs and indexes are at the bottom of pg_dump.
 				sqlDB.CheckQueryResults(t, `SHOW CREATE TABLE second`, [][]string{{
 					"second", `CREATE TABLE second (
-	i INT NOT NULL,
+	i INT8 NOT NULL,
 	s STRING NULL,
 	CONSTRAINT second_pkey PRIMARY KEY (i ASC),
 	FAMILY "primary" (i, s)
@@ -2367,8 +2367,8 @@ func TestImportPgDump(t *testing.T) {
 			if c.expected == expectAll {
 				sqlDB.CheckQueryResults(t, `SHOW CREATE TABLE seqtable`, [][]string{{
 					"seqtable", `CREATE TABLE seqtable (
-	a INT NULL DEFAULT nextval('public.a_seq':::STRING),
-	b INT NULL,
+	a INT8 NULL DEFAULT nextval('public.a_seq':::STRING),
+	b INT8 NULL,
 	FAMILY "primary" (a, b, rowid)
 )`,
 				}})
@@ -2424,7 +2424,7 @@ func TestImportCockroachDump(t *testing.T) {
 	})
 	sqlDB.CheckQueryResults(t, "SHOW CREATE TABLE t", [][]string{
 		{"t", `CREATE TABLE t (
-	i INT NOT NULL,
+	i INT8 NOT NULL,
 	t STRING NULL,
 	CONSTRAINT "primary" PRIMARY KEY (i ASC),
 	INDEX t_t_idx (t ASC),
@@ -2433,7 +2433,7 @@ func TestImportCockroachDump(t *testing.T) {
 	})
 	sqlDB.CheckQueryResults(t, "SHOW CREATE TABLE a", [][]string{
 		{"a", `CREATE TABLE a (
-	i INT NOT NULL,
+	i INT8 NOT NULL,
 	CONSTRAINT "primary" PRIMARY KEY (i ASC),
 	CONSTRAINT fk_i_ref_t FOREIGN KEY (i) REFERENCES t (i),
 	FAMILY "primary" (i)
